improve search for newer information of LibreOffice Help Pack
We now sort by name in descending order instead of by date.
This should work as long as the version number stays below 10.x.
I know that this might not be accurate, but since e.g. version
5.2.6 was released/updated after version 5.3.0, the old approach
with the modification date might have delivered an older version
(a.k.a "LibreOffice Still" instead of "LibreOffice Fresh"), and
that is not the newest stable version.

I need to rework that as soon as LibreOffice 10 comes out.