major refactoring for options; add detection for PDF24 Creator
The refactoring was necessary to keep the main program class as
simple as possible and to make sure that future options can be
easily added to the program without the need to rewrite all major
classes of the application.

Command line options are now stored in a separate class, and the
parsing of command line arguments is handled by another new class
as well. Separation of concerns helps to keep things clear, nice
and clean.