fix possible problem with checksums for CDBurnerXP
When searching for newer versions of CDBurnerXP, the updater
cannot get any checksums for the installers, because the official
CDBurnerXP website does not provide any. To avoid that at least
when the newest version is identical to the newest version known
to updater, the updater will fall back to the checksums of the
known version. That way we have proper checksums.