avoid possible problem with HTTPS certificate on CCleaner website
The CCleaner website is delivered via a CDN, so the site uses a
certificate for the CDN domain and not for the real domain of the
developer (piriform.com). Therefore, web requests to the site may
fail, if they use HTTPS instead of HTTP. To avoid that, the web
request in the searchForNewer() method for CCleaner will use HTTP
and no HTTPS anymore. That should fix things.

However, the installer will still be downloaded via a HTTPS
connection, so we should still be save on the transport level.
Additionally, the installer is a signed executable, and that
signature is checked before installation.