implement predefined / known checksums for Firefox Aurora
Additionally, Firefox Developer Edition (aurora) is now always
enabled and part of the software list, not just when explicitly
specified via command line parameter. This can be done, because
there is no performance penalty for Aurora anymore, since
checksums are already there and must not be retrieved for all
Aurora versions during the program run.