add workaround for VLC mirror servers with invalid certificates
The canonical download URL for VLC's checksum files, i. e.
<https://get.videolan.org/vlc/3.0.8/win32/vlc-3.0.8-win32.exe.sha256>
randomly redirect to one of the various VLC mirror servers. Some
of them do not seem to have a proper TLS certificate (or maybe it
is alright, but Windows just does not have it in its trust store).
So to work around this issue, the program tries another mirror
server instead. This is no compromise in security, because all
such requests are directed to mirrors anyway, so we might as well
pick one of them directly.