add expiration date to signature data of installers
That way the signature is only tested when the certificate has
not expired yet. It should avoid some problems due to outdated
signature / publisher information. However, it may also lead to
some installers not being verifiable anymore, if the signature
is the only way for verification. In that case the update will
not be performed.