allow pre-update processes to fail for Inkscape
Since there seems to be no better way to ensure that all old MSI
installations of Inkscape are purged than trying to uninstall all
known older Inkscape MSI setups, we need to allow some of them to
fail, because (hopefully) not all old Inkscape versions are still
installed. Usually only one of the MSI uninstalls will succeed
and all others will fail. That is alright in this very specific
case.