adjust log level for some messages
The "Searching for newer version" type messages are now info
level messages and not debug level messages anymore. Furthermore,
the remaining debug level messages are only logged to the console
target, if the binary was build in debug mode. In release mode
only info level messages (or higher log levels) are printed to
the console target.