update known information for Thunderbird to version 91.0
Since both the Farsi and Sinhalese language versions of
Thunderbird have been removed in version 91.0, there is now a
separate class that takes care of these, allowing those language
versions to be updated to version 78.13.0. While this is not an
optimal solution it is probably the best thing to do as there is
no clear migration option for those languages.