update known information for Eclipse Temurin JDK 8 to version 8u302-b08
This is the first release of the AdoptOpenJDK JDK 8 that is
distributed by the Eclipse Foundation as Eclipse Temurin after
the transition of AdoptOpenJDK to the Eclipse Foundation. It is
now known as the Eclipse Adoptium project. More on that change
can be read on
<https://blog.adoptopenjdk.net/2021/03/transition-to-eclipse-an-update/>.
Furthermore, Eclipse has rebranded AdoptOpenJDK. Hence the new
name.

The default installation location has also changed from
C:\Program Files\AdoptOpenJDK\jdk-<version>-hotspot to
C:\Program Files\Eclipse Foundation\jdk-<version>-hotspot.