update known information for Eclipse Temurin JRE 11 to version 11.0.14.1+1
This is the current release of the AdoptOpenJDK JRE 11 that is
distributed by the Eclipse Foundation as Eclipse Temurin after
the transition of AdoptOpenJDK to the Eclipse Foundation. It is
now known as the Eclipse Adoptium project. More on that change
can be read on
<https://blog.adoptopenjdk.net/2021/03/transition-to-eclipse-an-update/>.
Furthermore, Eclipse has rebranded AdoptOpenJDK. Hence the new
name.

The default installation location has also changed from
C:\Program Files\AdoptOpenJDK\jre-<version>-hotspot to
C:\Program Files\Eclipse Adoptium\jre-<version>-hotspot.