switch from WebClient to HttpClient in moz-checksum-generator
WebClient is obsolete, starting with .NET 6, and HttpClient is
the recommended replacement for it. I'm not sure whether this is
really changing anything for the better, because on some
platforms HttpClient uses the same backend as WebClient did to
perform the actual requests. But at least we can get rid of some
of those warnings.