change network code to reuse HttpClient when possible
Basically, System.Net.Http.HttpClient was introduced with the
intention to be reusable. So whenever a standard request without
any "fancy" modifications is required, it can be reused. However,
I still am not sure how this instance gets cleaned up in the end,
because no explicit .Dispose() is called and the GC is not always
guaranteed to clean this up, as far as I understand.