prepare for 64-bit-only builds of Mumble client
Starting with version 1.5.x, the Mumble developers do no longer
provide 32 bit builds of the client. This means that 32 bit OSes
have to stay at version 1.4.287. However, on 64 bit OSes that
still use the 32 bit version of the Mumble client, a switch to
the 64 bit version of Mumble can be performed to allow further
updates to happen. Machines that were using the 64 bit version
from the start can get updates as usual.