decode contentwrapper element on FileZilla's website
The FileZilla website wraps the actual version and download links
in a Base64-encoded and AES-CBC-encrypted text that gets decrypted
via JavaScript in the browser. However, the updater does not run
JavaScript code when checking the FileZilla website, so the same
algorithm to decode the website content has to be built using the
C# tools available.